requirements.txt
fastapi
uvicorn[standard]
python-dotenv
requests
transformers
torch
sentence-transformers
google-genai
streamlit 
folium 
streamlit-folium
geopy